node-proxy: rename server.py -> main.py
This is going to be the entrypoint of node-proxy, let's rename
this file to main.py
